requirements.txt
# Placeholder for requirements.txt

flask==2.0.1
face-recognition==1.3.0
opencv-python==4.5.3.56
numpy==1.21.2
pillow==8.3.2
